+++ nyalib-base/src/main/java/net/danygames2014/nyalib/fluid/FluidHandlerItem.java
@@ -11,6 +11,7 @@ public interface FluidHandlerItem {
      * Check if the entity supports extracting fluid if this returns false there
      * should be no point in trying to use {@link #extractFluid(ItemStack, int, int)}
      *
+     * @param thiz The ItemStack this is called on
      * @return <code>true</code> if the handler supports fluid extraction
      */
     boolean canExtractFluid(ItemStack thiz);
@@ -18,6 +19,7 @@ public interface FluidHandlerItem {
     /**
      * Extract a fluid in the given slot from the entity
      *
+     * @param thiz The ItemStack this is called on
      * @param slot      The slot to extract from
      * @param amount    The amount in mB to extract
      * @return The FluidStack extracted, null if nothing is extracted
@@ -27,6 +29,7 @@ public interface FluidHandlerItem {
     /**
      * Extract any fluid from the entity
      *
+     * @param thiz The ItemStack this is called on
      * @return The extracted {@link FluidStack}
      */
     default FluidStack extractFluid(ItemStack thiz) {
@@ -36,6 +39,7 @@ public interface FluidHandlerItem {
     /**
      * Extract a specified amount of any fluid from the entity
      *
+     * @param thiz The ItemStack this is called on
      * @param amount    The amount of fluid in mB to extract
      * @return The extracted {@link FluidStack}
      */
@@ -51,6 +55,7 @@ public interface FluidHandlerItem {
     /**
      * Extract the given fluid in any slot from the handler
      *
+     * @param thiz The ItemStack this is called on
      * @param fluid     The Fluid to extract
      * @param amount    The amount in mB to extract
      * @return The FluidStack extracted, null if nothing is extracted
@@ -84,6 +89,7 @@ public interface FluidHandlerItem {
      * Check if the entity supports inserting fluids, if this returns false there
      * should be no point in trying to use {@link #insertFluid(ItemStack, FluidStack)} or {@link #insertFluid(ItemStack, FluidStack, int)}
      *
+     * @param thiz The ItemStack this is called on
      * @return <code>true</code> if the entity supports fluid insertion
      */
     boolean canInsertFluid(ItemStack thiz);
@@ -91,6 +97,7 @@ public interface FluidHandlerItem {
     /**
      * Insert fluid into the given slot and return the remainder
      *
+     * @param thiz The ItemStack this is called on
      * @param stack     The {@link FluidStack} to insert
      * @param slot      Slot to insert into
      * @return The remainder of the FluidStack (null if it was inserted entirely), this should be a new FluidStack, however it can be the same if it was not modified
@@ -100,6 +107,7 @@ public interface FluidHandlerItem {
     /**
      * Insert fluid into any slot and return the remainder
      *
+     * @param thiz The ItemStack this is called on
      * @param stack     The {@link FluidStack} to insert
      * @return The remainder of the FluidStack (null if it was inserted entirely), this should be a new FluidStack, however it can be the same if it was not modified
      */
@@ -109,6 +117,7 @@ public interface FluidHandlerItem {
      * Get the {@link FluidStack} in the given slot, If there is no {@link FluidStack}, then return null
      * <p>
      *
+     * @param thiz The ItemStack this is called on
      * @param slot      The slot to get the {@link FluidStack} from
      * @return The {@link FluidStack} in the slot
      */
@@ -117,6 +126,7 @@ public interface FluidHandlerItem {
     /**
      * Sets a {@link FluidStack} into the given slot
      *
+     * @param thiz The ItemStack this is called on
      * @param slot      The slot to set the {@link FluidStack} into
      * @param stack     The {@link FluidStack} to set into the slot
      * @return Whether the action was succesfull
@@ -126,6 +136,7 @@ public interface FluidHandlerItem {
     /**
      * Get the size of the entity fluid inventory
      *
+     * @param thiz The ItemStack this is called on
      * @return The number of slots this entity has
      */
     int getFluidSlots(ItemStack thiz);
@@ -133,6 +144,7 @@ public interface FluidHandlerItem {
     /**
      * Get the capacity of the given slot
      *
+     * @param thiz The ItemStack this is called on
      * @param slot      The slot to query for capacity
      * @return The capacity of the slot
      */
@@ -141,6 +153,7 @@ public interface FluidHandlerItem {
     /**
      * Get the entire fluid inventory of the entity
      *
+     * @param thiz The ItemStack this is called on
      * @return An array of all the FluidStacks
      */
     FluidStack[] getFluids(ItemStack thiz);